EXPERIMENTAL: Bugfix for style net + BREAK keyword (and >75 tokens).
The context is received in two separate forward calls, so the extra_cond
is appended to each of them with the assumption that each of (pos,neg)
will appear intermittently, and considering ddim/plms will reverse the
order of appearance, as I've observed.

NOTE: I have been informed that UNIPC sampler has a similar reversal,
should consider adding it to the vanilla list.