Update batch_hijack.py
change the comparison method of InputMode 

I'm modifying the webui code to use controlnet+adetailer as rest api, I found that this code causes problem. 

The problem is the class variable of InputMode is instantiated for each thread (UI and Rest API), so the comparison from Rest API doesn't work properly. In Python, it is not regarded as string, so the comparison is done with their id.  so I changed the comparison code into just simple string comparison.